python: merge python-kadmin-rs into kadmin (#61)
