+++ .github/workflows/ci-python.yml
@@ -50,7 +50,6 @@ jobs:
       fail-fast: false
       matrix:
         python-version:
-          - "3.8"
           - "3.9"
           - "3.10"
           - "3.11"
@@ -235,7 +234,7 @@ jobs:
     environment: pypi
     permissions:
       id-token: write
-    if: github.event_name == 'release' && startsWith(github.ref, 'refs/tags/python-kadmin-rs/version/')
+    if: github.event_name == 'release' && startsWith(github.ref, 'refs/tags/kadmin/version/')
     steps:
       - uses: actions/download-artifact@v4
         with:
