+++ .github/workflows/release-pr.yml
@@ -11,7 +11,6 @@ on:
         options:
           - kadmin-sys
           - kadmin
-          - python-kadmin-rs
       version:
         description: Version to release
         required: true
