+++ .readthedocs.yaml
@@ -15,12 +15,12 @@ build:
 
 python:
   install:
-    - requirements: python-kadmin-rs/docs/requirements.txt
+    - requirements: python/docs/requirements.txt
     - method: pip
       path: .
 
-# Build documentation in the "docs/" directory with Sphinx
+# Build documentation in the "python/docs/" directory with Sphinx
 sphinx:
   builder: html
-  configuration: python-kadmin-rs/docs/conf.py
+  configuration: python/docs/conf.py
   fail_on_warning: true
