+++ kadmin/Cargo.toml
@@ -16,11 +16,18 @@ categories = ["api-bindings", "authentication"]
 default = ["client"]
 client = ["kadmin-sys/client"]
 local = ["kadmin-sys/server"]
+python = ["dep:pyo3", "dep:indoc"]
 
 [dependencies]
 chrono = "0.4"
 getset = "0.1"
+indoc = { version = "2", optional = true }
 kadmin-sys = { path = "../kadmin-sys", version = "0.2.0", default-features = false }
+pyo3 = { version = "0.23", features = [
+  "extension-module",
+  "chrono",
+  "chrono-tz",
+], optional = true }
 thiserror = "2"
 
 [dev-dependencies]
@@ -34,3 +41,8 @@ serial_test = { version = "3.1", default-features = false, features = [
 
 [lints]
 workspace = true
+
+[package.metadata.release]
+pre-release-replacements = [
+  { file = "../pyproject.toml", search = "^version = \"[a-z0-9\\.-]+\"", replace = "version = \"{{version}}\"", exactly = 1 },
+]
