+++ kadmin/src/db_args.rs
@@ -2,6 +2,9 @@
 
 use std::{ffi::CString, os::raw::c_char, ptr::null_mut};
 
+#[cfg(feature = "python")]
+use pyo3::prelude::*;
+
 use crate::error::Result;
 
 /// Database specific arguments
@@ -15,6 +18,7 @@ use crate::error::Result;
 ///     .unwrap();
 /// ```
 #[derive(Debug)]
+#[cfg_attr(feature = "python", pyclass)]
 pub struct DbArgs {
     /// NULL-terminated list of strings of the form `arg[=value]`
     ///
