--- python-kadmin-rs/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "python-kadmin-rs"
-version = "0.3.0"
-edition.workspace = true
-rust-version.workspace = true
-
-[features]
-default = ["client"]
-client = ["kadmin/client"]
-local = ["kadmin/local"]
-
-[dependencies]
-indoc = "2"
-kadmin = { path = "../kadmin", version = "0.3.0", default-features = false }
-pyo3 = { version = "0.23", features = [
-  "extension-module",
-  "chrono",
-  "chrono-tz",
-] }
-
-[lints]
-workspace = true
-
-[package.metadata.release]
-pre-release-replacements = [
-  { file = "../pyproject.toml", search = "^version = \"[a-z0-9\\.-]+\"", replace = "version = \"{{version}}\"", exactly = 1 },
-]
