--- python-kadmin-rs/docs/kadmin.rst
@@ -1,15 +1,0 @@
-kadmin
-======
-
-.. automodule:: kadmin
-   :members:
-   :undoc-members:
-   :imported-members:
-
-Exceptions
-----------
-
-.. automodule:: kadmin.exceptions
-   :members:
-   :undoc-members:
-   :imported-members:
