--- python-kadmin-rs/docs/kadmin_local.rst
@@ -1,17 +1,0 @@
-kadmin_local
-============
-
-.. automodule:: kadmin_local
-   :members:
-   :undoc-members:
-   :imported-members:
-   :noindex:
-
-Exceptions
-----------
-
-.. automodule:: kadmin_local.exceptions
-   :members:
-   :undoc-members:
-   :imported-members:
-   :noindex:
