+++ python/docs/conf.py
@@ -5,7 +5,6 @@ copyright = "authentik community"
 
 extensions = [
     "sphinx.ext.autodoc",
-    "sphinx.ext.autosummary",
     "sphinx_rtd_theme",
 ]
 exclude_patterns = ["_build"]
