+++ python/src/kadmin_local/__init__.pyi
@@ -18,29 +18,8 @@ class KAdmin:
     def policy_exists(self, name: str) -> bool: ...
     def list_policies(self, query: str | None = None) -> List[str]: ...
     @staticmethod
-    def with_password(
-        client_name: str,
-        password: str,
-        params: Params | None = None,
-        db_args: DbArgs | None = None,
-    ) -> KAdmin: ...
-    @staticmethod
-    def with_keytab(
-        client_name: str | None = None,
-        keytab: str | None = None,
-        params: Params | None = None,
-        db_args: DbArgs | None = None,
-    ) -> KAdmin: ...
-    @staticmethod
-    def with_ccache(
-        client_name: str | None = None,
-        ccache_name: str | None = None,
-        params: Params | None = None,
-        db_args: DbArgs | None = None,
-    ) -> KAdmin: ...
-    @staticmethod
-    def with_anonymous(
-        client_name: str, params: Params | None = None, db_args: DbArgs | None = None
+    def with_local(
+        params: Params | None = None, db_args: DbArgs | None = None
     ) -> KAdmin: ...
 
 @final
@@ -60,18 +39,47 @@ class Policy:
     max_renewable_life: datetime.timedelta | None
     tl_data: TlData
 
-    def modify(self, **kwargs) -> Policy: ...
-    def delete(self) -> None: ...
+    def modify(self, kadmin: KAdmin, **kwargs) -> Policy: ...
+    def delete(self, kadmin: KAdmin) -> None: ...
 
 @final
 class Principal:
-    def change_password(self, password: str): ...
+    name: str
+    expire_time: datetime.datetime | None
+    last_password_change: datetime.datetime | None
+    password_expiration: datetime.datetime | None
+    max_life: datetime.timedelta | None
+    modified_by: str
+    modified_at: datetime.datetime | None
+    attributes: int
+    kvno: int
+    mkvno: int
+    policy: str | None
+    aux_attributes: int
+    max_renewable_life: datetime.timedelta | None
+    last_success: datetime.datetime | None
+    last_failed: datetime.datetime | None
+    fail_auth_count: int
+
+    def change_password(self, kadmin: KAdmin, password: str): ...
 
 @final
-class Params: ...
+class Params:
+    def __init__(
+        self,
+        realm: str | None = None,
+        kadmind_port: int | None = None,
+        kpasswd_port: int | None = None,
+        admin_server: str | None = None,
+        dbname: str | None = None,
+        acl_file: str | None = None,
+        dict_file: str | None = None,
+        stash_file: str | None = None,
+    ): ...
 
 @final
-class DbArgs: ...
+class DbArgs:
+    def __init__(self, /, *args: str, **kwargs: str | None): ...
 
 @final
 class TlDataEntry:
